Allow manual running of reviewdog on full repo

+++ assets/reviewdog.sh
@@ -6,7 +6,7 @@ export SCRIPTPATH=`dirname $SCRIPT`
 export GOPATH=$HOME/go
 export PATH=$PATH:$GOROOT/bin:$GOPATH/bin
 
-RUNNERS="safesvg tfsec semgrep brakeman npm-audit pip-audit sveltegrep"
+RUNNERS="safesvg tfsec semgrep sveltegrep brakeman npm-audit pip-audit"
 
 if [ -n "${GITHUB_BASE_REF+set}" ]; then
     for runner in $RUNNERS; do
@@ -19,9 +19,22 @@ if [ -n "${GITHUB_BASE_REF+set}" ]; then
         cat $runner.log >> reviewdog.log
         wc -l $runner.log
     done
+
 else
-    find $SCRIPTPATH/../t3sts/ | sed "s|$SCRIPTPATH/../||g" | tr '\n' '\0' > $SCRIPTPATH/all_changed_files.txt
-    GITHUB_BASE_REF=initial-commit reviewdog  -runners=semgrep,safesvg,sveltegrep,pip-audit,npm-audit -conf="$SCRIPTPATH/reviewdog/reviewdog.yml"  -diff="git diff origin/$GITHUB_BASE_REF" -reporter=local -tee
+    git ls-files | tr '\n' '\0' > $SCRIPTPATH/all_changed_files.txt
+    reviewdog \
+      -runners=$(echo "$RUNNERS" | tr ' ' ',') \
+      -conf="$SCRIPTPATH/reviewdog/reviewdog.yml"  \
+      -filter-mode=nofilter \
+      -reporter=local \
+      -tee \
+      | sed 's/<br><br>Cc @brave\/sec-team[ ]*//' \
+      | tee reviewdog.log
+fi
+
+FAIL=$(cat reviewdog.log | grep 'failed with zero findings: The command itself failed' || true)
+if [[ -n "$FAIL" ]]; then
+    exit 101
 fi
 
 find reviewdog.log -type f -empty -delete
